Change File from having filePath to having path, name and directory
Previously, a File's path was stored under a property named `filePath`.

Now, File has two string properties, `name` and `directory` which store
the name and directory of the File respectively. In addition to this,
there is also a dynamic `path` property which when read, will resolve
to a string mimicking the old `filePath` property. When set, it will
update the underlying `name` and `directory` properties.